continue geometry refactoring

+++ src/opencv/opencv.h
@@ -52,7 +52,7 @@ void cvMultH(Homography H1, Homography H2, Homography *Hout);
 void cvInvertH(Homography *Hom);
 void cvApplyFlips(Homography *Hom, int source_ry, int target_ry);
 
-void cvGetMatrixReframe(int x, int y, int w, int h, double angle, Homography *Hom);
+void cvGetMatrixReframe(double x, double y, int w, int h, double angle, Homography *Hom);
 void cvGetBoundingRectSize(fits *image, point center, double angle, int *w, int *h);
 
 #ifdef __cplusplus
